+++ app/src/main/java/de/tu_chemnitz/tomkr/augmentedmaps/sensor/LocationService.java
@@ -39,7 +39,7 @@ public class LocationService {
 
     public void pushLocation(){
         for(de.tu_chemnitz.tomkr.augmentedmaps.sensor.LocationListener listener : listeners){
-            listener.onInitialLocation(new de.tu_chemnitz.tomkr.augmentedmaps.core.basetypes.Location((float)lastBigLocation.getLatitude(), (float)lastBigLocation.getLongitude(), (float)lastBigLocation.getAltitude()));
+            listener.onInitialLocation(new de.tu_chemnitz.tomkr.augmentedmaps.core.basetypes.Location((float)lastBigLocation.getLatitude(), (float)lastBigLocation.getLongitude(), (int)lastBigLocation.getAltitude()));
         }
     }
 
@@ -108,11 +108,11 @@ public class LocationService {
             if(lastBigLocation.distanceTo(location)>DIST){
                 lastBigLocation = location;
                 for(de.tu_chemnitz.tomkr.augmentedmaps.sensor.LocationListener listener : listeners){
-                    listener.onBigLocationChange(new de.tu_chemnitz.tomkr.augmentedmaps.core.basetypes.Location((float)location.getLatitude(), (float)location.getLongitude(), (float)location.getAltitude()));
+                    listener.onBigLocationChange(new de.tu_chemnitz.tomkr.augmentedmaps.core.basetypes.Location((float)location.getLatitude(), (float)location.getLongitude(), (int)location.getAltitude()));
                 }
             }
             for(de.tu_chemnitz.tomkr.augmentedmaps.sensor.LocationListener listener : listeners){
-                listener.onSmallLocationChange(new de.tu_chemnitz.tomkr.augmentedmaps.core.basetypes.Location((float)location.getLatitude(), (float)location.getLongitude(), (float)location.getAltitude()));
+                listener.onSmallLocationChange(new de.tu_chemnitz.tomkr.augmentedmaps.core.basetypes.Location((float)location.getLatitude(), (float)location.getLongitude(), (int)location.getAltitude()));
             }
         }
